pyproject.toml
[project]
name = 'openmm_csvr'
version = '1.0.0'
dependencies=['numpy','openmm', 'pymbar-core', 'scipy', 'mdtraj', 'matplotlib']

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["openmm_csvr*"]

[tool.black]
line-length = 88
exclude = '''
\(
    \.git
  | build
\)
'''

[tool.isort]
profile = "black"
